+++ src/java/com/github/jsdossier/proto/Dossier.java
@@ -151,83 +151,84 @@ public final class Dossier {
       "\030\003 \001(\t\"\202\001\n\007Comment\022%\n\005token\030\001 \003(\0132\026.doss",
       "ier.Comment.Token\032P\n\005Token\022\016\n\004text\030\001 \001(\t" +
       "H\000\022\026\n\004html\030\006 \001(\tB\006\362\201\031\002\020\001H\000\022\024\n\004href\030\004 \001(\t" +
-      "B\006\362\201\031\002\010\001B\t\n\007content\"\242\001\n\004Tags\022\020\n\010is_const" +
+      "B\006\362\201\031\002\010\001B\t\n\007content\"\265\001\n\004Tags\022\020\n\010is_const" +
       "\030\001 \001(\010\022\025\n\ris_deprecated\030\002 \001(\010\022\017\n\007is_dict" +
       "\030\003 \001(\010\022\020\n\010is_final\030\004 \001(\010\022\024\n\014is_interface" +
       "\030\005 \001(\010\022\021\n\tis_module\030\006 \001(\010\022\021\n\tis_struct\030\007" +
-      " \001(\010\022\022\n\nis_default\030\010 \001(\010\"\364\002\n\014BasePropert" +
-      "y\022\014\n\004name\030\001 \002(\t\022#\n\006source\030\002 \002(\0132\023.dossie" +
-      "r.SourceLink\022%\n\013description\030\003 \002(\0132\020.doss" +
-      "ier.Comment\022%\n\013deprecation\030\004 \001(\0132\020.dossi",
-      "er.Comment\022/\n\nvisibility\030\005 \001(\0162\023.dossier" +
-      ".Visibility:\006PUBLIC\022\033\n\004tags\030\006 \001(\0132\r.doss" +
-      "ier.Tags\022$\n\ndefined_by\030\007 \001(\0132\020.dossier.C" +
-      "omment\022#\n\toverrides\030\010 \001(\0132\020.dossier.Comm" +
-      "ent\022&\n\014specified_by\030\t \003(\0132\020.dossier.Comm" +
-      "ent\022\"\n\010see_also\030\n \003(\0132\020.dossier.Comment\"" +
-      "O\n\010Property\022#\n\004base\030\001 \002(\0132\025.dossier.Base" +
-      "Property\022\036\n\004type\030\002 \001(\0132\020.dossier.Comment" +
-      "\"\305\002\n\010Function\022#\n\004base\030\001 \002(\0132\025.dossier.Ba" +
-      "seProperty\022\025\n\rtemplate_name\030\002 \003(\t\022\035\n\016is_",
-      "constructor\030\003 \001(\010:\005false\022+\n\tparameter\030\004 " +
-      "\003(\0132\030.dossier.Function.Detail\022(\n\006return\030" +
-      "\005 \001(\0132\030.dossier.Function.Detail\022(\n\006throw" +
-      "n\030\006 \003(\0132\030.dossier.Function.Detail\032]\n\006Det" +
-      "ail\022\014\n\004name\030\001 \001(\t\022\036\n\004type\030\002 \001(\0132\020.dossie" +
-      "r.Comment\022%\n\013description\030\003 \001(\0132\020.dossier" +
-      ".Comment\"\356\001\n\013Enumeration\022\036\n\004type\030\001 \002(\0132\020" +
-      ".dossier.Comment\022)\n\005value\030\002 \003(\0132\032.dossie" +
-      "r.Enumeration.Value\022/\n\nvisibility\030\003 \001(\0162" +
-      "\023.dossier.Visibility:\006PUBLIC\032c\n\005Value\022\014\n",
-      "\004name\030\001 \002(\t\022%\n\013description\030\002 \001(\0132\020.dossi" +
-      "er.Comment\022%\n\013deprecation\030\003 \001(\0132\020.dossie" +
-      "r.Comment\"o\n\005Index\022\024\n\004home\030\001 \002(\tB\006\362\201\031\002\010\001" +
-      "\022\025\n\rinclude_types\030\002 \001(\010\022\027\n\017include_modul" +
-      "es\030\003 \001(\010\022 \n\005links\030\004 \003(\0132\021.dossier.TypeLi" +
-      "nk\"\226\t\n\006JsType\022\014\n\004name\030\001 \002(\t\022#\n\006source\030\005 " +
-      "\002(\0132\023.dossier.SourceLink\022+\n\006nested\030\006 \003(\013" +
-      "2\033.dossier.JsType.TypeSummary\022%\n\013descrip" +
-      "tion\030\007 \002(\0132\020.dossier.Comment\022\033\n\004tags\030\010 \002" +
-      "(\0132\r.dossier.Tags\022%\n\013deprecation\030\t \001(\0132\020",
-      ".dossier.Comment\022)\n\010type_def\030\n \003(\0132\027.dos" +
-      "sier.JsType.TypeDef\022)\n\013enumeration\030\013 \001(\013" +
-      "2\024.dossier.Enumeration\022*\n\017static_functio" +
-      "n\030\014 \003(\0132\021.dossier.Function\022*\n\017static_pro" +
-      "perty\030\r \003(\0132\021.dossier.Property\022(\n\rmain_f" +
-      "unction\030\016 \001(\0132\021.dossier.Function\022!\n\006meth" +
-      "od\030\017 \003(\0132\021.dossier.Function\022 \n\005field\030\020 \003" +
-      "(\0132\021.dossier.Property\022\'\n\rextended_type\030\022" +
-      " \003(\0132\020.dossier.Comment\022*\n\020implemented_ty" +
-      "pe\030\023 \003(\0132\020.dossier.Comment\022,\n\021compiler_c",
-      "onstant\030\024 \003(\0132\021.dossier.Property\022*\n\006pare" +
-      "nt\030\025 \001(\0132\032.dossier.JsType.ParentLink\022\'\n\014" +
-      "aliased_type\030\026 \001(\0132\021.dossier.TypeLink\022\020\n" +
-      "\010filename\030\027 \001(\t\022\026\n\016qualified_name\030\030 \001(\t\022" +
-      ",\n\021reexported_module\030\031 \003(\0132\021.dossier.Pro" +
-      "perty\032T\n\013TypeSummary\022\014\n\004name\030\001 \002(\t\022\024\n\004hr" +
-      "ef\030\002 \002(\tB\006\362\201\031\002\010\001\022!\n\007summary\030\003 \002(\0132\020.doss" +
-      "ier.Comment\032\333\001\n\007TypeDef\022\014\n\004name\030\001 \002(\t\022\036\n" +
-      "\004type\030\002 \002(\0132\020.dossier.Comment\022#\n\006source\030" +
-      "\003 \002(\0132\023.dossier.SourceLink\022%\n\013descriptio",
-      "n\030\004 \002(\0132\020.dossier.Comment\022%\n\013deprecation" +
-      "\030\005 \001(\0132\020.dossier.Comment\022/\n\nvisibility\030\006" +
-      " \001(\0162\023.dossier.Visibility:\006PUBLIC\032@\n\nPar" +
-      "entLink\022\037\n\004link\030\001 \002(\0132\021.dossier.TypeLink" +
-      "\022\021\n\tis_module\030\002 \001(\010\"\210\001\n\016HtmlRenderSpec\022%" +
-      "\n\tresources\030\001 \002(\0132\022.dossier.Resources\022\r\n" +
-      "\005title\030\002 \002(\t\022!\n\007content\030\003 \001(\0132\020.dossier." +
-      "Comment\022\035\n\005index\030\004 \002(\0132\016.dossier.Index\"\177" +
-      "\n\024SourceFileRenderSpec\022%\n\tresources\030\001 \002(" +
-      "\0132\022.dossier.Resources\022!\n\004file\030\002 \002(\0132\023.do",
-      "ssier.SourceFile\022\035\n\005index\030\003 \002(\0132\016.dossie" +
-      "r.Index\"w\n\020JsTypeRenderSpec\022\035\n\004type\030\001 \003(" +
-      "\0132\017.dossier.JsType\022%\n\tresources\030\002 \002(\0132\022." +
-      "dossier.Resources\022\035\n\005index\030\003 \002(\0132\016.dossi" +
-      "er.Index*A\n\nVisibility\022\n\n\006PUBLIC\020\000\022\r\n\tPR" +
-      "OTECTED\020\001\022\013\n\007PRIVATE\020\002\022\013\n\007PACKAGE\020\003:M\n\ts" +
-      "anitized\022\035.google.protobuf.FieldOptions\030" +
-      "\236\220\003 \001(\0132\031.dossier.SanitizedContentB\036\n\032co" +
-      "m.github.jsdossier.protoP\001"
+      " \001(\010\022\022\n\nis_default\030\010 \001(\010\022\021\n\tis_record\030\t " +
+      "\001(\010\"\364\002\n\014BaseProperty\022\014\n\004name\030\001 \002(\t\022#\n\006so" +
+      "urce\030\002 \002(\0132\023.dossier.SourceLink\022%\n\013descr" +
+      "iption\030\003 \002(\0132\020.dossier.Comment\022%\n\013deprec",
+      "ation\030\004 \001(\0132\020.dossier.Comment\022/\n\nvisibil" +
+      "ity\030\005 \001(\0162\023.dossier.Visibility:\006PUBLIC\022\033" +
+      "\n\004tags\030\006 \001(\0132\r.dossier.Tags\022$\n\ndefined_b" +
+      "y\030\007 \001(\0132\020.dossier.Comment\022#\n\toverrides\030\010" +
+      " \001(\0132\020.dossier.Comment\022&\n\014specified_by\030\t" +
+      " \003(\0132\020.dossier.Comment\022\"\n\010see_also\030\n \003(\013" +
+      "2\020.dossier.Comment\"O\n\010Property\022#\n\004base\030\001" +
+      " \002(\0132\025.dossier.BaseProperty\022\036\n\004type\030\002 \001(" +
+      "\0132\020.dossier.Comment\"\305\002\n\010Function\022#\n\004base" +
+      "\030\001 \002(\0132\025.dossier.BaseProperty\022\025\n\rtemplat",
+      "e_name\030\002 \003(\t\022\035\n\016is_constructor\030\003 \001(\010:\005fa" +
+      "lse\022+\n\tparameter\030\004 \003(\0132\030.dossier.Functio" +
+      "n.Detail\022(\n\006return\030\005 \001(\0132\030.dossier.Funct" +
+      "ion.Detail\022(\n\006thrown\030\006 \003(\0132\030.dossier.Fun" +
+      "ction.Detail\032]\n\006Detail\022\014\n\004name\030\001 \001(\t\022\036\n\004" +
+      "type\030\002 \001(\0132\020.dossier.Comment\022%\n\013descript" +
+      "ion\030\003 \001(\0132\020.dossier.Comment\"\356\001\n\013Enumerat" +
+      "ion\022\036\n\004type\030\001 \002(\0132\020.dossier.Comment\022)\n\005v" +
+      "alue\030\002 \003(\0132\032.dossier.Enumeration.Value\022/" +
+      "\n\nvisibility\030\003 \001(\0162\023.dossier.Visibility:",
+      "\006PUBLIC\032c\n\005Value\022\014\n\004name\030\001 \002(\t\022%\n\013descri" +
+      "ption\030\002 \001(\0132\020.dossier.Comment\022%\n\013depreca" +
+      "tion\030\003 \001(\0132\020.dossier.Comment\"o\n\005Index\022\024\n" +
+      "\004home\030\001 \002(\tB\006\362\201\031\002\010\001\022\025\n\rinclude_types\030\002 \001" +
+      "(\010\022\027\n\017include_modules\030\003 \001(\010\022 \n\005links\030\004 \003" +
+      "(\0132\021.dossier.TypeLink\"\226\t\n\006JsType\022\014\n\004name" +
+      "\030\001 \002(\t\022#\n\006source\030\005 \002(\0132\023.dossier.SourceL" +
+      "ink\022+\n\006nested\030\006 \003(\0132\033.dossier.JsType.Typ" +
+      "eSummary\022%\n\013description\030\007 \002(\0132\020.dossier." +
+      "Comment\022\033\n\004tags\030\010 \002(\0132\r.dossier.Tags\022%\n\013",
+      "deprecation\030\t \001(\0132\020.dossier.Comment\022)\n\010t" +
+      "ype_def\030\n \003(\0132\027.dossier.JsType.TypeDef\022)" +
+      "\n\013enumeration\030\013 \001(\0132\024.dossier.Enumeratio" +
+      "n\022*\n\017static_function\030\014 \003(\0132\021.dossier.Fun" +
+      "ction\022*\n\017static_property\030\r \003(\0132\021.dossier" +
+      ".Property\022(\n\rmain_function\030\016 \001(\0132\021.dossi" +
+      "er.Function\022!\n\006method\030\017 \003(\0132\021.dossier.Fu" +
+      "nction\022 \n\005field\030\020 \003(\0132\021.dossier.Property" +
+      "\022\'\n\rextended_type\030\022 \003(\0132\020.dossier.Commen" +
+      "t\022*\n\020implemented_type\030\023 \003(\0132\020.dossier.Co",
+      "mment\022,\n\021compiler_constant\030\024 \003(\0132\021.dossi" +
+      "er.Property\022*\n\006parent\030\025 \001(\0132\032.dossier.Js" +
+      "Type.ParentLink\022\'\n\014aliased_type\030\026 \001(\0132\021." +
+      "dossier.TypeLink\022\020\n\010filename\030\027 \001(\t\022\026\n\016qu" +
+      "alified_name\030\030 \001(\t\022,\n\021reexported_module\030" +
+      "\031 \003(\0132\021.dossier.Property\032T\n\013TypeSummary\022" +
+      "\014\n\004name\030\001 \002(\t\022\024\n\004href\030\002 \002(\tB\006\362\201\031\002\010\001\022!\n\007s" +
+      "ummary\030\003 \002(\0132\020.dossier.Comment\032\333\001\n\007TypeD" +
+      "ef\022\014\n\004name\030\001 \002(\t\022\036\n\004type\030\002 \002(\0132\020.dossier" +
+      ".Comment\022#\n\006source\030\003 \002(\0132\023.dossier.Sourc",
+      "eLink\022%\n\013description\030\004 \002(\0132\020.dossier.Com" +
+      "ment\022%\n\013deprecation\030\005 \001(\0132\020.dossier.Comm" +
+      "ent\022/\n\nvisibility\030\006 \001(\0162\023.dossier.Visibi" +
+      "lity:\006PUBLIC\032@\n\nParentLink\022\037\n\004link\030\001 \002(\013" +
+      "2\021.dossier.TypeLink\022\021\n\tis_module\030\002 \001(\010\"\210" +
+      "\001\n\016HtmlRenderSpec\022%\n\tresources\030\001 \002(\0132\022.d" +
+      "ossier.Resources\022\r\n\005title\030\002 \002(\t\022!\n\007conte" +
+      "nt\030\003 \001(\0132\020.dossier.Comment\022\035\n\005index\030\004 \002(" +
+      "\0132\016.dossier.Index\"\177\n\024SourceFileRenderSpe" +
+      "c\022%\n\tresources\030\001 \002(\0132\022.dossier.Resources",
+      "\022!\n\004file\030\002 \002(\0132\023.dossier.SourceFile\022\035\n\005i" +
+      "ndex\030\003 \002(\0132\016.dossier.Index\"w\n\020JsTypeRend" +
+      "erSpec\022\035\n\004type\030\001 \003(\0132\017.dossier.JsType\022%\n" +
+      "\tresources\030\002 \002(\0132\022.dossier.Resources\022\035\n\005" +
+      "index\030\003 \002(\0132\016.dossier.Index*A\n\nVisibilit" +
+      "y\022\n\n\006PUBLIC\020\000\022\r\n\tPROTECTED\020\001\022\013\n\007PRIVATE\020" +
+      "\002\022\013\n\007PACKAGE\020\003:M\n\tsanitized\022\035.google.pro" +
+      "tobuf.FieldOptions\030\236\220\003 \001(\0132\031.dossier.San" +
+      "itizedContentB\036\n\032com.github.jsdossier.pr" +
+      "otoP\001"
     };
     com.google.protobuf.Descriptors.FileDescriptor.InternalDescriptorAssigner assigner =
         new com.google.protobuf.Descriptors.FileDescriptor.    InternalDescriptorAssigner() {
@@ -289,7 +290,7 @@ public final class Dossier {
     internal_static_dossier_Tags_fieldAccessorTable = new
       com.google.protobuf.GeneratedMessage.FieldAccessorTable(
         internal_static_dossier_Tags_descriptor,
-        new java.lang.String[] { "IsConst", "IsDeprecated", "IsDict", "IsFinal", "IsInterface", "IsModule", "IsStruct", "IsDefault", });
+        new java.lang.String[] { "IsConst", "IsDeprecated", "IsDict", "IsFinal", "IsInterface", "IsModule", "IsStruct", "IsDefault", "IsRecord", });
     internal_static_dossier_BaseProperty_descriptor =
       getDescriptor().getMessageTypes().get(7);
     internal_static_dossier_BaseProperty_fieldAccessorTable = new